Fixing evaluation of expressions
For nested operator expressions, the evaluation would evaluate the nodes
bottom up and would put the value for a node at the same place in a
tree isomorphic to the expression tree. However, when evaluating
some operator, the operands actually used would always be the immediate
child nodes of the root of the value tree and not the child nodes of the
node corresponding to the operator node.
This patch fixes that.